Add initial integration tests for ed25519
Based on the raft integration test, they are
mainly the same tests, but executed in a network created with
ed25519 crypto material.

Note that in this commit we did not added the ed25519 integration
test network creation, implemented in the "integration/nwo" folder.
The present commit only add the tests.

Signed-off-by: Johann Westphall <johannwestphall@gmail.com>